+++ lib/src/private/sub_extension/context/index.dart
@@ -4,4 +4,5 @@ export 'context_duration_extension.dart';
 export 'context_general_extension.dart';
 export 'context_navigator_extension.dart';
 export 'context_padding_extension.dart';
+export 'context_popup_extension.dart';
 export 'context_size_extension.dart';